Add an optional kernel_size to gaussian kernel
The gaussian kernel interpolation is necessarily slower than the
bilinear or nearest neighbour interpolations as it takes into
account the pixels over the whole image.

Thus, add a kerenl_size parameter that can be set by the user in
order to vary the size of the kernel, to allow for faster
computations.